requirements.txt
Flask==2.3.2
Pillow==10.0.0
exif==1.3.5
